Move ppu funcs from pst_from to pp_utils
